Made refactoring. Errors.go now will be used by all other structures as response, so it was moved to pkg folder.
