+++ internal/goals/mock_goal_repository.go
@@ -1,5 +1,9 @@
 package goals
 
+import (
+	"github.com/hvmello/goal-tracker-backend/pkg/response"
+)
+
 // MockRepository implements Repository interface for testing
 type MockRepository struct {
 	goals map[uint]Goal
@@ -24,12 +28,12 @@ func (m *MockRepository) FindByID(id uint) (*Goal, error) {
 	if goal, exists := m.goals[id]; exists {
 		return &goal, nil
 	}
-	return nil, ErrGoalNotFound
+	return nil, response.ErrGoalNotFound
 }
 
 func (m *MockRepository) Create(goal *Goal) error {
 	if goal.ID == 0 {
-		return ErrInvalidID
+		return response.ErrInvalidID
 	}
 	m.goals[goal.ID] = *goal
 	return nil
@@ -37,7 +41,7 @@ func (m *MockRepository) Create(goal *Goal) error {
 
 func (m *MockRepository) Update(goal *Goal) error {
 	if _, exists := m.goals[goal.ID]; !exists {
-		return ErrGoalNotFound
+		return response.ErrGoalNotFound
 	}
 	m.goals[goal.ID] = *goal
 	return nil
@@ -45,7 +49,7 @@ func (m *MockRepository) Update(goal *Goal) error {
 
 func (m *MockRepository) Delete(id uint) error {
 	if _, exists := m.goals[id]; !exists {
-		return ErrGoalNotFound
+		return response.ErrGoalNotFound
 	}
 	delete(m.goals, id)
 	return nil
