+++ pkg/response/errors.go
@@ -1,4 +1,4 @@
-package goals
+package response
 
 import (
 	"net/http"
